Removed Compatibility for 16.42.3 & Older
This change really shouldn’t of happened but it did because of my dumbness. I brought this back so that Modern Versions of YouTube only works on YTABConfig. The reason for it is dumb but I’m eliminating support for legacy versions because those will eventually stop working so I’ve had to do this.
and PoomSmart. Sorry that I’ve been a jerk. I was stupid for what I’ve said.