Updated Readme
I haven’t been liking these new changes you done in commit “dad60dd” so I’m taking actions on making YTABConfig to work with only v17.30.1 or newer versions of YouTube like it was before. Adding support to v16.42.3 was so stupid and isn’t a good idea. Hopefully you can reverse it because it was me that should’ve never gave you that idea.