Removed Compatibility for 16.42.3 & Older
This change really shouldn’t have happened. But then again it happened because of my dumbness. So I brought this code back so that Modern Versions of YouTube only works on YTABConfig. The reason for it is dumb and reasonable but I’m eliminating support for legacy versions because those versions are gone!
and PoomSmart. Sorry that I’ve been a jerk. I was stupid for what I’ve said and sorry that you changed YTABConfig. Hopefully this will commit bring back the old YTABConfig as far as I know.