Removed Compatibility for 16.42.3 & Older
This change really shouldn't have happened. But then again it happened because of that stupid discussion I made. So since I’m sorry, I decided to bring back this method. So that modern versions of YouTube like v17.33.2+ will only work on YTABConfig. The reason for it is dumb and reasonable, but I'm eliminating support for legacy versions because those versions like v16.xx.x being gone! And PoomSmart, once again, I am sorry that for being a jerk. I was stupid for what I've said in some of the discussion and sorry it effected you to changing YTABConfig. I hope this commit somehow eventually reverses the weird implementation you’ve done in the future. and once again. I am sorry.