Failsafe + removed v16.42.3 support
Added a decently suggestive feature that PoomSmart surprisingly declined. Adding a failsafe will hope a ton of people fix when enabling certain stuff in YTABConfig Settings. Restart the app 5 times under 60 seconds will cause the app to forcefully and trigger the “Reset and Kill” Option from YTABConfig settings. Which resets any option you changed in A/B, also as a bonus, this only works with Modern Versions of the YouTube App because I despised some of the code PoomSmart had.