Added Failsafe Implementation
This was a very sugguested feature from PoomSmart’s YTABConfig tweak which was definitely a needed addition for YTABConfig whenever the user accidentally enabled an option that can potentially be app-breaking, locked out, crashes & etc. so I am doing what I can to help PoomSmart help with this code. It is definitely one of the best implementation I’ve done so far for PoomSmart. Enjoy.